Fix build on Windows

+++ interpreter/src/api.cpp
@@ -54,7 +54,9 @@ namespace snek::interpreter::api
    static value::ptr
    Exit(Runtime&, const std::vector<value::ptr>& arguments)
    {
-     std::exit(static_cast<const value::Int*>(arguments[0].get())->value);
+     std::exit(static_cast<int>(
+       static_cast<const value::Int*>(arguments[0].get())->value)
+     );
 
      return nullptr;
    }
